requirements.txt
altgraph==0.17
colorama==0.4.3
cycler==0.10.0
future==0.18.2
kiwisolver==1.2.0
logzero==1.5.0
matplotlib==3.0.0
numpy==1.18.5
pampy==0.3.0
pandas==0.25.3
pefile==2019.4.18
prompt-toolkit==3.0.5
PyInstaller==3.6
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
pywin32-ctypes==0.2.0
PyYAML==5.3.1
scipy==1.4.1
seaborn==0.10.1
six==1.15.0
wcwidth==0.2.4
yaspin==0.17.0
